text input doesn't get stored in command line history
line_reader has no affordance for checking history, so rather
than write tests for this fix, I validated manually. There's
probably a reasonable implementation of command line history
and the LineRead interface and/or LineReader library that
would allow for it, but there are higher priority tasks, so
exploring that will have to wait.